Split MCWalkerConfiguration
